requirements.txt
spotipy
instagrapi
python-dotenv
pillow